requirements.txt
fastapi
uvicorn
motor
google-generativeai
python-dotenv
requests
google-api-python-client
